requirements-dev.txt
black
mypy
pre-commit
pytest
pytest-flask
